initial attempt to support translations of the website through the 'alternatives' feature, including initial German translation as a test case (of course, only the index page is actually translated)
This was a little bit clumsier than I'd hoped (see e.g.
https://github.com/lektor/lektor/issues/275#issuecomment-457612662) but
it still seems like the best way to handle this

Includes support for a mechanism I came up with to allow
language-specific extensions to the basic site template; this will be
used for example by the French translation of the site.